PruebaBackendDeveloper/src/co/edu/usbbog/ada/pruebaBackendDeveloper/modelo/Producto.java
package co.edu.usbbog.ada.pruebaBackendDeveloper.modelo;

/**
 *
 * @author SantiagoEcheDuque
 */
public class Producto {

    private int id;
    private String codigo;
    private String nombre;
    private String marca;
    private String fechaVenci; // fecha de vencimiento
    private double costo;
    private int cantidad;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getCodigo() {
        return codigo;
    }

    public void setCodigo(String codigo) {
        this.codigo = codigo;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public String getMarca() {
        return marca;
    }

    public void setMarca(String marca) {
        this.marca = marca;
    }

    public String getFechaVenci() {
        return fechaVenci;
    }

    public void setFechaVenci(String fechaVenci) {
        this.fechaVenci = fechaVenci;
    }

    public double getCosto() {
        return costo;
    }

    public void setCosto(double costo) {
        this.costo = costo;
    }

    public int getCantidad() {
        return cantidad;
    }

    public void setCantidad(int cantidad) {
        this.cantidad = cantidad;
    }

    @Override
    public String toString() {
        return "Producto{" + "codigo=" + codigo + ", nombre=" + nombre + ", marca=" + marca + ", fechaVenci=" + fechaVenci + ", costo=" + costo + ", cantidad=" + cantidad + '}';
    }

    
    
}
